requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.24.3
mysql-connector-python==8.2.0
requests==2.31.0
lxml==4.9.3
sqlalchemy==2.0.41
pymysql==1.1.0
xgboost==2.0.3
scikit-learn==1.7.1